Implement transactionality when buying books.

+++ src/main/java/ru/dbpractice/learnup/db/model/Book.java
@@ -21,7 +21,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "book_id")
-    private Long book_id;
+    private int book_id;
 
     @Column
     private String title;
